type fixes to diff_drive_kinematics

+++ firmware/libraries/diffDriveKinematics/diff_drive_kinematics.cpp
@@ -9,10 +9,10 @@ Robot::Robot(int pulses_per_revolution, float wheel_radius, float distance_betwe
 
 void Robot::WheelDistance(unsigned long int wheel_ticks_left, unsigned long int wheel_ticks_right, float &leftdistance, float &rightdistance)
 {
-    int wheel_circumference = wheel_radius_ * M_PI * 2;
+    float wheel_circumference = wheel_radius_ * M_PI * 2;
 
-    leftdistance = (wheel_ticks_left * wheel_circumference) / pulses_per_revolution_;
-    rightdistance = (wheel_ticks_right * wheel_circumference) / pulses_per_revolution_;
+    leftdistance = (wheel_ticks_left * wheel_circumference) / float(pulses_per_revolution_);
+    rightdistance = (wheel_ticks_right * wheel_circumference) / float(pulses_per_revolution_);
 }
 
 void Robot::ComputePose(float left_distance, float right_distance, float &tetha, float &rx, float &ry)
